html/browsers/history/the-location-interface/location-protocol-setter.html

<!doctype html>
<title>Set location.protocol to broken schemes</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id=log></div>
<iframe src="data:text/html,<script>
onmessage = (e) => {
  let results = [];
  e.data.forEach((val) => {
    try {
      location.protocol = val;
      results.push('failure')
    } catch(e) {
      results.push(e.name)
    }
  });
  parent.postMessage(results, '*')
}
</script>"></iframe>
<iframe srcdoc="<script>
onmessage = (e) => {
  let results = [];
  e.data.forEach((val) => {
    try {
      location.protocol = val;
      results.push('failure')
    } catch(e) {
      results.push(e.name)
    }
  });
  parent.postMessage(results, '*')
}
</script>"></iframe>
<script>
  let broken = [
    'x\x20',
    'x\x00',
    '\x00',
    ':',
    '†',
    'a†',
    'a\x0A',
    'a+-.\x0A',
    '\x7F',
    '\xFF'
  ]
  ;broken.forEach((val) => {
    test(() => {
      assert_throws(new TypeError, () => { location.protocol = val })
    }, encodeURI(val) + " (percent-encoded) is not a scheme")
  })
  let c = 0
  async_test((t) => {
    self.onload = t.step_func(() => {
      self.onmessage = t.step_func((e) => {
        assert_array_equals(e.data, ('TypeError '.repeat(9) + 'TypeError').split(' '))
        c++
        if(c === 2) {
          t.done()
        }
      })
      self[0].postMessage(broken, "*")
      self[1].postMessage(broken, "*")
    })
  }, "Equivalent tests for data URL and srcdoc <iframe>s")
</script>
